Day 6 part 1 refactor + day 6 part 2 soulution

+++ src/day6/DataProvider.java
@@ -5,9 +5,9 @@ import java.util.List;
 public interface DataProvider {
 
 
-    List<Integer> provideTimes();
+    List<Long> provideTimes();
 
-    List<Integer> provideDistances();
+    List<Long> provideDistances();
 
 }
